# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram RPG Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Player {
  id: number;
  telegram_id: string;
  username: string;
  level: number;
  exp: number;
  exp_to_next: number;
  gold: number;
  /** @nullable */
  hero_type: string | null;
  hero_level: number;
}

export type HeroSelectionHeroType = typeof HeroSelectionHeroType[keyof typeof HeroSelectionHeroType];


export const HeroSelectionHeroType = {
  warrior: 'warrior',
  mage: 'mage',
  assassin: 'assassin',
} as const;

export interface HeroSelection {
  hero_type: HeroSelectionHeroType;
}

export type BattleInputResult = typeof BattleInputResult[keyof typeof BattleInputResult];


export const BattleInputResult = {
  win: 'win',
  lose: 'lose',
} as const;

export interface BattleInput {
  telegram_id: string;
  monster_name: string;
  result: BattleInputResult;
  exp_gained: number;
  gold_gained: number;
}

export interface Battle {
  id: number;
  player_id: number;
  monster_name: string;
  result: string;
  exp_gained: number;
  gold_gained: number;
  created_at: string;
}

export interface BattleResult {
  battle: Battle;
  player: Player;
}

export interface LeaderboardEntry {
  rank: number;
  username: string;
  level: number;
  gold: number;
  /** @nullable */
  hero_type: string | null;
}

export type GetPlayerParams = {
telegram_id: string;
username?: string;
};

export type SelectHeroParams = {
telegram_id: string;
};

export type UpgradeSkillParams = {
telegram_id: string;
};

export type GetBattleHistoryParams = {
telegram_id: string;
limit?: number;
};

export type GetLeaderboardParams = {
limit?: number;
};