# lib/api-zod/src/generated/types/battle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram RPG Game API
 * OpenAPI spec version: 0.1.0
 */

export interface Battle {
  id: number;
  player_id: number;
  monster_name: string;
  result: string;
  exp_gained: number;
  gold_gained: number;
  created_at: string;
}