# lib/api-zod/src/generated/types/battleInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram RPG Game API
 * OpenAPI spec version: 0.1.0
 */
import type { BattleInputResult } from './battleInputResult';

export interface BattleInput {
  telegram_id: string;
  monster_name: string;
  result: BattleInputResult;
  exp_gained: number;
  gold_gained: number;
}