# lib/api-zod/src/generated/types/battleResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram RPG Game API
 * OpenAPI spec version: 0.1.0
 */
import type { Battle } from './battle';
import type { Player } from './player';

export interface BattleResult {
  battle: Battle;
  player: Player;
}