# lib/api-zod/src/generated/types/heroSelection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram RPG Game API
 * OpenAPI spec version: 0.1.0
 */
import type { HeroSelectionHeroType } from './heroSelectionHeroType';

export interface HeroSelection {
  hero_type: HeroSelectionHeroType;
}