# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram RPG Game API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  rank: number;
  username: string;
  level: number;
  gold: number;
  /** @nullable */
  hero_type: string | null;
}