# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram RPG Game API
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: number;
  telegram_id: string;
  username: string;
  level: number;
  exp: number;
  exp_to_next: number;
  gold: number;
  /** @nullable */
  hero_type: string | null;
  hero_level: number;
}